Add setup mode

+++ run-tt-rss.sh
@@ -1,25 +1,47 @@
 #!/bin/sh
 
+# Path to tt-rss instance
+ttrssPath='/usr/share/nginx/tt-rss/'
+
 # Path to tt-rss feed updater to use. Either update.php --daemon or update_daemon2.php can be used.
-updater='/usr/share/nginx/tt-rss/update_daemon2.php'
+updater="$ttrssPath/update_daemon2.php"
 
 # Start the web server
 service php5-fpm start
 service nginx start
 
-# Run updater as ww-data user
-su -s '/bin/sh' -c "$updater" www-data &
-updaterPID=$!
+if [ "$SETUP" = 'true' ]; then
+	echo "RUNNING IN SETUP MODE!" 1>&2
 
-# Check that all services are running every 10 seconds, and if not kill the whole container.
-while service php5-fpm status >/dev/null 2>&1 && \
-	service nginx status >/dev/null 2>&1 && \
-	kill -0 "$updaterPID" >/dev/null 2>&1; do
+	# If we're using the nginx proxy, output a url for the setup address
+	if [ -n "$VIRTUAL_HOST" ]; then
+		echo "Configre tt-rss at http://$VIRTUAL_HOST/install/" 1>&2
+	fi
 
-	sleep 10
-done
+	# Run until config file is generated
+	while ! [ -e "$ttrssPath/config.php" ]; do
+		sleep 1
+	done
 
-# Stop all processes before exiting
-kill "$updaterPID"
+	echo "Config file found, goodbye!" 1>&2
+else
+	# Run updater as ww-data user
+	su -s '/bin/sh' -c "$updater" www-data &
+	updaterPID=$!
+
+	# Check that all services are running every 10 seconds, and if not kill the whole container.
+	while service php5-fpm status >/dev/null 2>&1 && \
+		service nginx status >/dev/null 2>&1 && \
+		kill -0 "$updaterPID" >/dev/null 2>&1; do
+
+		sleep 10
+	done
+
+	# Stop updater
+	kill "$updaterPID"
+
+fi
+
+# Stop webserver
 service nginx stop
 service php5-fpm stop
